requirements.txt
streamlit>=1.37
pandas>=2.2
sqlalchemy>=2.0
python-dotenv>=1.0
psycopg[binary]>=3.2

